requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6
jinja2==3.1.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
aiofiles==0.24.0
pillow==10.1.0
reportlab==4.0.7
qrcode==7.4.2
websockets==12.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
gunicorn==21.2.0
pymongo==4.13.0
uuid6==2024.7.10
flask==3.0.0
python-dateutil==2.9.0.post0
tenacity==8.5.0 